refactor: update htmlStorageLocation filename
