+++ apps/web/src/app/core/html-storage.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 
 import { AbstractStorageService } from "@bitwarden/common/abstractions/storage.service";
-import { HtmlStorageLocation } from "@bitwarden/common/enums/htmlStorageLocation";
+import { HtmlStorageLocation } from "@bitwarden/common/enums/html-storage-location";
 import { StorageOptions } from "@bitwarden/common/models/domain/storage-options";
 
 @Injectable()
