+++ libs/common/src/enums/html-storage-location.ts
