+++ libs/common/src/services/state.service.ts
@@ -17,7 +17,7 @@ import { Policy } from "../admin-console/models/domain/policy";
 import { CollectionView } from "../admin-console/models/view/collection.view";
 import { EnvironmentUrls } from "../auth/models/domain/environment-urls";
 import { KdfConfig } from "../auth/models/domain/kdf-config";
-import { HtmlStorageLocation } from "../enums/htmlStorageLocation";
+import { HtmlStorageLocation } from "../enums/html-storage-location";
 import { KdfType } from "../enums/kdfType";
 import { StorageLocation } from "../enums/storageLocation";
 import { ThemeType } from "../enums/themeType";
